backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pika
pillow